Replace Conditional with Polymorphism
Create subclasses matching the branches of the conditional.
In them, create a shared method and move code from the corresponding
branch of the conditional to it. Then replace the conditional with
the relevant method call. The result is that the proper implementation
will be attained via polymorphism depending on the object class.